[WIP] Well Known Hashing
Password is very sensative information and need to be stored as secure
as possible:
hashing is one of the well known ways to do that.
But because of some techniques (such as rainbow table) its not enough
to use only hash function. For that reson 'salt' is aslo used.

Made steps of security:

- every user will have randomly genereted salt based on password which
  is passed by him/her (passed password will be shuffled and hashed
  to create new salt),
- passed in password also will be hashed,
- salt and hashed password will be cocatenated.

Later This salt will be stored in database with user other credentials
(password, username, etc.):
to check whether passed in password is correct or not server dose the
following:
- finds row of user credentials by comparing passed in username (unique) with
  username feald of specified table,
- takes salt which was stored as discussed above,
- cocatenates passed in password with salt and hashes them,
- compares new hash with stored password.