Files state is separated from ImageUpload controller
To include iamge to table it is required to know image file name, but
because of previouse definition it can't be used:
$scope  object can't be accessed from parent, because of inheritance and
prototype principles.
To share some state between controllers/$scopes angular service is used:
now instead of holding files state in $scope object it will be maintained in
'FilesStateManipulator' service and to get fiels state it just requires to
call 'getFilesState' method!